requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
amqp==5.1.1
    # via
    #   -r requirements.in
    #   kombu
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
bleach==6.0.0
    # via readme-renderer
build==0.10.0
    # via pip-tools
celery==5.2.7
    # via
    #   -r requirements.in
    #   flower
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-log
    #   click-plugins
    #   click-repl
    #   nltk
    #   pip-tools
    #   python-semantic-release
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-log==0.4.0
    # via python-semantic-release
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.6
    # via twine
cryptography==40.0.2
    # via secretstorage
deprecation==2.1.0
    # via postgrest
docutils==0.19
    # via readme-renderer
dotty-dict==1.3.1
    # via python-semantic-release
fastapi==0.95.1
    # via -r requirements.in
flower==1.2.0
    # via -r requirements.in
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via python-semantic-release
gotrue==1.0.1
    # via supabase
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.3
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.23.3
    # via
    #   gotrue
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
humanize==4.6.0
    # via flower
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
importlib-metadata==6.4.1
    # via
    #   keyring
    #   twine
invoke==1.7.3
    # via python-semantic-release
jaraco-classes==3.2.3
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
joblib==1.2.0
    # via nltk
keyring==23.13.1
    # via twine
kombu==5.2.4
    # via celery
more-itertools==9.1.0
    # via jaraco-classes
nltk==3.8.1
    # via -r requirements.in
packaging==23.1
    # via
    #   build
    #   deprecation
    #   python-semantic-release
pip-tools==6.13.0
    # via -r requirements.in
pkginfo==1.9.6
    # via twine
postgrest==0.10.6
    # via supabase
prometheus-client==0.16.0
    # via flower
prompt-toolkit==3.0.38
    # via click-repl
psycopg2==2.9.6
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==1.10.7
    # via
    #   fastapi
    #   gotrue
    #   postgrest
pygments==2.15.0
    # via readme-renderer
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via
    #   realtime
    #   storage3
python-dotenv==1.0.0
    # via uvicorn
python-gitlab==3.14.0
    # via python-semantic-release
python-semantic-release==7.33.2
    # via supabase
pytz==2023.3
    # via
    #   celery
    #   flower
pyyaml==6.0
    # via uvicorn
readme-renderer==37.3
    # via twine
realtime==1.0.0
    # via supabase
redis==4.5.4
    # via -r requirements.in
regex==2023.3.23
    # via nltk
requests==2.28.2
    # via
    #   python-gitlab
    #   python-semantic-release
    #   requests-toolbelt
    #   twine
requests-toolbelt==0.10.1
    # via
    #   python-gitlab
    #   twine
rfc3986[idna2008]==1.5.0
    # via
    #   httpx
    #   twine
secretstorage==3.3.3
    # via keyring
semver==2.13.0
    # via python-semantic-release
six==1.16.0
    # via
    #   bleach
    #   click-repl
    #   python-dateutil
smmap==5.0.0
    # via gitdb
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.26.1
    # via fastapi
storage3==0.5.2
    # via supabase
strenum==0.4.10
    # via postgrest
supabase==1.0.3
    # via -r requirements.in
supafunc==0.2.2
    # via supabase
tomli==2.0.1
    # via
    #   build
    #   pyproject-hooks
tomlkit==0.11.7
    # via python-semantic-release
tornado==6.2
    # via flower
tqdm==4.65.0
    # via
    #   nltk
    #   twine
twine==3.8.0
    # via python-semantic-release
typing-extensions==4.5.0
    # via
    #   pydantic
    #   realtime
    #   starlette
    #   storage3
urllib3==1.26.15
    # via
    #   requests
    #   twine
uvicorn[standard]==0.21.1
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.19.0
    # via uvicorn
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
websockets==10.4
    # via
    #   realtime
    #   uvicorn
wheel==0.40.0
    # via
    #   pip-tools
    #   python-semantic-release
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
